TableView - fixed a crash
When having a TableView on a window that was not the main window, if the TableView was scrolled down so that the first rows would not be visible anymore, and the user happened to close the window with the TableView, the program would crash.
The crash is because in QtTableView, the method _refresh_visible_row(self) gets called at least once after the user has closed the window, and at that time self.declaration is None, which implies a crash when accessing attributes visible_rows and visible_row.